Updated documentation and added additional checks in CartItem handlers

+++ Core/Ecommerce.Application/Features/CartItem/Commands/DeleteUserCartItems/DeleteUserCartItemsCommandHandler.cs
@@ -1,3 +1,4 @@
+using System;
 using Ecommerce.Persistence.Contracts;
 using Ecommerce.Shared.Responses.CartItem;
 using MediatR;
@@ -35,7 +36,7 @@ namespace Ecommerce.Application.Features.CartItem.Commands.DeleteUserCartItems
 		/// <returns>
 		/// A <see cref="DeleteUserCartItemsResponse"/> with Success being <c>true</c> if the <see cref="CartItem"/> entities were deleted;
 		/// Success will be <c>false</c> if no <see cref="CartItem"/> entities are found or validation of the command fails.
-		/// Message will contain the message to display.
+		/// Message will contain the message to display to the user.
 		/// </returns>
 		public async Task<DeleteUserCartItemsResponse> Handle(DeleteUserCartItemsCommand command, CancellationToken cancellationToken)
 		{
@@ -44,6 +45,14 @@ namespace Ecommerce.Application.Features.CartItem.Commands.DeleteUserCartItems
 			
 			//Create the response object
 			DeleteUserCartItemsResponse response = new DeleteUserCartItemsResponse { Success = true, Message = CartItemConstants._deleteAllItemsSuccessMessage };
+			
+			//Check if the UserId is valid
+			if (command.UserId == Guid.Empty)
+			{
+				response.Success = false;
+				response.Message = CartItemConstants._deleteAllItemsErrorMessage;
+				return response;
+			}
 			
 			//Attempt the delete
 			bool success = await this._cartItemRepository.RemoveUserCartItems(command.UserId);
